refactor(tools): migrate data tools to directory structure
- Refactored CSV tools (4 tools): csv-parser, csv-generator, csv-to-json, json-to-csv
- Refactored JSON tools (5 tools): json-parser, json-stringify, json-query, json-validator, json-merge
- Refactored XML tools (4 tools): xml-parser, xml-generator, xml-to-json, json-to-xml
- Refactored Transformer tools (6 tools): array-filter, array-map, array-sort, array-group-by, object-pick, object-omit

Each tool group now follows the established directory structure pattern:
- types.ts: Zod schemas and TypeScript interfaces
- index.ts: Default tool instances and factory functions
- tools/: Individual tool implementations (one per file)

Replaced monolithic files:
- csv-parser.ts (170 lines) -> csv/ directory
- json-processor.ts (149 lines) -> json/ directory
- xml-parser.ts (170 lines) -> xml/ directory
- transformer.ts (203 lines) -> transformer/ directory

All 975 tests passing
Build successful
No breaking changes to public API

Part of Phase 4 of the tools refactoring plan.
